Exports

RANSAC.FittedShape — Type

An abstract type that wraps the fitted shapes.

RANSAC.PointCloud — Method

PointCloud(vertices, normals, subsets, isenabled)

Construct a PointCloud with filling it's size and levelweight fields.

RANSAC.PointCloud — Method

PointCloud(vertices, normals, subsets, isenabled, size, levelweight, levelscore)

Constructor that converts vertices and normals to array of SVector{3,Float64}.

RANSAC.PointCloud — Method

PointCloud(vertices, normals, subsets)

Construct a PointCloud, filling it's other fields.

Every point is enabled, and the weight vector defaults to [1.0].

RANSAC.PointCloud — Method

PointCloud(vertices, normals, numofsubsets::Int)

Construct a PointCloud with numofsubsets random subsets.

Every point is enabled, and the weight vector defaults to [1.0].

RANSAC.E — Method

E(x::ConfidenceInterval)

Expected value of a ConfidenceInterval.

RANSAC.arbitrary_orthogonal — Method

arbitrary_orthogonal(vec)

Create an arbitrary orthogonal vector to vec.

RANSAC.arbitrary_orthogonal2 — Method

arbitrary_orthogonal2(vec)

Create an arbitrary orthogonal vector to vec.

RANSAC.dn2contour — Method

dn2contour(point, contour)

Distance: distance from the nearest point of the contour point.

RANSAC.dn2shape_contour — Method

dn2shape_contour(point, shape)

Distance: distance from the nearest point of the shape.contour point. Use in RANSAC.

RANSAC.dn2shape_outw — Method

dn2shape_outw(point, shape)

Distance: distance from the nearest point of the shape.contour point. Normal points always outwards. Use in CSG.

RANSAC.examplepc2 — Function

examplepc2(nois = false; all = false, mrotdeg = 10, vertscale = 0.5)

Generate a definitely not random example consisting planes and spheres.

RANSAC.examplepc3 — Function

examplepc3(nois = false; all = false, mrotdeg = 10, vertscale = 0.5)

Generate a definitely not random example consisting planes and spheres.

RANSAC.examplepc6 — Function

examplepc6(nois = false; all = false, mrotdeg = 10, vertscale = 0.5)

Generate a definitely not random example consisting planes and spheres.

RANSAC.findAABB — Method

findAABB(points)

Find the axis aligned minimum bounding box of the given pointcloud.

RANSAC.findhighestscore — Method

findhighestscore(A)

Find the largest expected value in an array of ScoredShapes.

Indicate if there's an overlap.

RANSAC.getcellandparents — Method

getcellandparents(cell::Cell)

Collect the cell and it's parents into an array.

RANSAC.havesameelement — Method

havesameelement(A, B)

Return true if A and B have at least one common element.

RANSAC.isoverlap — Method

isoverlap(i1::A, i2::A) where {A<:ConfidenceInterval}

Test if the two ConfidenceIntervals are overlapping.

RANSAC.isparallel — Method

isparallel(v1, v2, alpharad)

Check if v1 and v2 are parallel in an alpharad angle and point towards the same direction.

The vectors considered to be normalized.

RANSAC.iswithinrectangle — Method

iswithinrectangle(rect::HyperRectangle, p)

Decide if p (3 dimensional point) is within rect.

Works only with those prisms which axes are aligned with the coordinate frame. "Bottom/left" gives false and "top/right" gives true.

RANSAC.largestshape — Method

largestshape(A)

Find the largest shape in an array of ScoredShapes.

RANSAC.makemeanexample — Function

makemeanexample(nois = false; all = false)

Generate a definitely not random example.

RANSAC.noisifynormals — Method

noisifynormals(norms, maxrot)

Add gaussian noise to normals. Rotates the normals around a random axis with maxrot degrees.

Arguments:

maxrot::Real: maximum rotation in degrees.

RANSAC.noisifyvertices — Function

noisifyvertices(verts, allvs, scalef = 1)

Add gaussian noise to vertices. Random subset or all vertices can be selected.

Arguments:

allvs::Bool: adds noise to every vertice if true.

scalef::Real: scale the noise from the [-1,1] interval.

RANSAC.normalsforplot — Function

normalsforplot(verts, norms, arrowsize = 0.5)

Create an array of pair of points for plotting the normals with Makie. Only the direction of the normals is presented, their size not.

Arguments:

arrowsize: scaling factor for the size of the lines.

RANSAC.notsoconfident — Method

notsoconfident(x, y)

Return with a new ConfidenceInterval parameterized correctly with x and y.

RANSAC.prob — Method

prob(n, s, N, k)

The probability of successful detection of a shape sized n, from a point cloud size of N, with k size of minimal sets and s draws.

Arguments:

n: size of the shape.

s: number of candidates that have been drawn.

N: size of the point cloud.

k: size of the minimal set required to define a shape candidate.

RANSAC.project2plane — Method

project2plane(plane, points)

Project points on to the plane.

Arguments:

'plane::FittedPlane': a plane.

'points::AbstractArray': an array of points (array-of-arrays).

RANSAC.refitcylinder — Method

refit(s, pc, ϵ, α)

Refit cylinder. Only s.inpoints is updated.

RANSAC.refitplane — Method

refit(s, pc, ϵ, α)

Refit plane. Only s.inpoints is updated.

RANSAC.refitsphere — Method

refit(s, pc, ϵ, α)

Refit sphere. Only s.inpoints is updated.

RANSAC.rerunleftover! — Method

rerunleftover!(pc, nofs, params, sofarextr; reset_rand=true)

For example: rerunleftover(pcr, 4, p, extr, reset_rand=true). It modifies the list of extracted candidates (sofarextr) and pointcloud.isenabled.

RANSAC.rodriguesdeg — Method

rodriguesdeg(nv, ϑ)

Create a rotation matrix from an axis (nv) and an angle (ϑ) in degrees.

RANSAC.rodriguesrad — Method

rodriguesrad(nv, ϑ)

Create a rotation matrix from an axis (nv) and an angle (ϑ) in radians.

RANSAC.samplecylinder — Method

samplecylinder(ax, vp, R, h, sizet)

Create a tuple of points on a cylinder and their normals based on an axis (ax), a nullpoint (vp), radius and height.

Arguments:

ax: axis of the cylinder.

vp: nullpoint of the cylinder.

R: radius of the cylinder.

h: height of the cylinder.

sizet: tuple containing the number of samples along the circumference and the height.

RANSAC.sampleplane — Method

sampleplane(vp, n, lengtht, sizet)

Create a tuple of points on a plane and their normals based on a point (vp) and a surface normal (n).

Arguments:

vp: a point of the plane.

n: surface normal.

sizet: tuple containing the number of samples along each side.

lengtht: tuple containing the length of the plane along each side.

RANSAC.sampleplane — Method

sampleplane(vp, v1, v2, lengtht, sizet)

Create a tuple of points on a plane and their normals based on a point (vp) and two vectors (v1, v2).

Arguments:

vp: a point of the plane.

v1, v2: two vectors that define a plane.

sizet: tuple containing the number of samples along each side.

lengtht: tuple containing the length of the plane along each side.

RANSAC.samplesphere — Method

samplesphere(cp, R, sizet)

Create a tuple of points on a sphere and their normals.

Arguments:

cp: centerpoint of sphere.

R: radius of the sphere.

sizet: tuple containing the number of samples along two great circles.

RANSAC.smallestdistance — Method

smallestdistance(points)

Find the smallest distance between the points.

RANSAC.unitdisk2square — Method

unitdisk2square(p)

Transform point on unit disk to unit sqaure.

Maps a unit circle to a unit sqaure [0,1]^2. Source: shirley1997 - A low distortion map between disk and square